+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_params.py
@@ -173,7 +173,7 @@ def load_arguments(self, _):
         c.argument('service_cidr')
         c.argument('vnet_subnet_id')
         c.argument('workspace_resource_id')
-        c.argument('skip_role_assignment', required=False, action='store_true')
+        c.argument('skip-subnet-role-assignment', required=False, action='store_true')
 
     with self.argument_context('aks disable-addons') as c:
         c.argument('addons', options_list=['--addons', '-a'])